Fix API connector modal layout
