+++ packages/server/src/api/controllers/query/import/sources/openapi3.ts
@@ -16,6 +16,7 @@ import {
   generateRequestBodyFromSchema,
   buildKeyValueRequestBody,
 } from "./utils/requestBody"
+import { buildEndpointName } from "./utils/endpointName"
 
 type ServerObject = OpenAPIV3.ServerObject
 type ServerVariableObject = OpenAPIV3.ServerVariableObject
@@ -47,182 +48,6 @@ const isOperation = (
   pathItem: any
 ): pathItem is OpenAPIV3.OperationObject => {
   return methods.includes(key)
-}
-
-const OPERATION_ID_MAX_LENGTH = 60
-const ACTION_SEGMENTS: Record<string, string> = {
-  search: "Search",
-  read: "Read",
-  create: "Create",
-  update: "Update",
-  upsert: "Upsert",
-  delete: "Delete",
-  archive: "Archive",
-  restore: "Restore",
-  merge: "Merge",
-}
-const METHOD_ACTIONS: Record<string, string> = {
-  get: "Get",
-  post: "Create",
-  put: "Update",
-  patch: "Update",
-  delete: "Delete",
-}
-const CONTEXT_SEGMENTS = new Set([
-  "account",
-  "auth",
-  "automation",
-  "business",
-  "communication",
-  "conversations",
-  "crm",
-  "cms",
-  "events",
-  "files",
-  "marketing",
-  "meta",
-  "scheduler",
-  "settings",
-  "webhooks",
-])
-const WORD_REPLACEMENTS: Record<string, string> = {
-  api: "API",
-  cms: "CMS",
-  crm: "CRM",
-  id: "ID",
-  oauth: "OAuth",
-  scim: "SCIM",
-  ui: "UI",
-  url: "URL",
-}
-
-const isSimpleOperationId = (operationId?: string): boolean => {
-  if (!operationId) {
-    return false
-  }
-  const trimmed = operationId.trim()
-  if (!trimmed) {
-    return false
-  }
-  if (trimmed.length > OPERATION_ID_MAX_LENGTH) {
-    return false
-  }
-  if (/[{}]/.test(trimmed)) {
-    return false
-  }
-  return !trimmed.includes("/")
-}
-
-const getShortText = (value?: string): string | undefined => {
-  if (!value) {
-    return undefined
-  }
-  const trimmed = value.trim()
-  if (!trimmed) {
-    return undefined
-  }
-  const firstLine = trimmed
-    .split(/\r?\n/)
-    .map(line => line.trim())
-    .find(line => line.length > 0)
-  if (!firstLine) {
-    return undefined
-  }
-  const maxLength = 80
-  if (firstLine.length > maxLength) {
-    return `${firstLine.slice(0, maxLength - 3).trim()}...`
-  }
-  return firstLine
-}
-
-const isVersionSegment = (segment: string): boolean => {
-  const lower = segment.toLowerCase()
-  return (
-    /^v\d+$/.test(lower) ||
-    /^\d{4}-\d{2}(-\d{2})?$/.test(lower) ||
-    /^\d+$/.test(lower)
-  )
-}
-
-const isPathParamSegment = (segment: string): boolean => {
-  return segment.startsWith("{") && segment.endsWith("}")
-}
-
-const humanizeSegment = (segment: string): string => {
-  const spaced = segment
-    .replace(/([a-z0-9])([A-Z])/g, "$1 $2")
-    .replace(/[-_]/g, " ")
-  return spaced
-    .split(/\s+/)
-    .filter(Boolean)
-    .map(word => {
-      const lower = word.toLowerCase()
-      const replacement = WORD_REPLACEMENTS[lower]
-      if (replacement) {
-        return replacement
-      }
-      return `${word[0].toUpperCase()}${word.slice(1)}`
-    })
-    .join(" ")
-}
-
-const buildResourceName = (segments: string[]): string => {
-  if (!segments.length) {
-    return ""
-  }
-  let selected = segments
-  if (segments.length > 2 && CONTEXT_SEGMENTS.has(segments[0].toLowerCase())) {
-    selected = [segments[0], ...segments.slice(-2)]
-  } else if (segments.length > 2) {
-    selected = segments.slice(-2)
-  }
-  return selected.map(humanizeSegment).join(" ")
-}
-
-const buildFallbackName = (methodName: string, path: string): string => {
-  const normalizedMethod = methodName.toLowerCase()
-  let segments = path.split("/").filter(Boolean)
-  segments = segments.filter(
-    segment => !isVersionSegment(segment) && !isPathParamSegment(segment)
-  )
-  const isBatch = segments.some(segment => segment.toLowerCase() === "batch")
-  segments = segments.filter(segment => segment.toLowerCase() !== "batch")
-  let action = METHOD_ACTIONS[normalizedMethod] || "Call"
-  const last = segments[segments.length - 1]
-  if (last) {
-    const actionLabel = ACTION_SEGMENTS[last.toLowerCase()]
-    if (actionLabel) {
-      action = actionLabel
-      segments = segments.slice(0, -1)
-    }
-  }
-  if (isBatch && action !== "Batch") {
-    action = `Batch ${action}`
-  }
-  const resource = buildResourceName(segments)
-  if (!resource) {
-    return action
-  }
-  return `${action} ${resource}`
-}
-
-const buildEndpointName = (
-  operation: OpenAPIV3.OperationObject,
-  methodName: string,
-  path: string
-): string => {
-  if (isSimpleOperationId(operation.operationId)) {
-    return operation.operationId?.trim() || path
-  }
-  const summary = getShortText(operation.summary)
-  if (summary) {
-    return summary
-  }
-  const description = getShortText(operation.description)
-  if (description) {
-    return description
-  }
-  return buildFallbackName(methodName, path)
 }
 
 /**
